requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langchain
langgraph
langchain-openai
langchain-community
bs4
faiss-cpu
python-dotenv
